Updates to the save message
I was having a rough time trying to implement the save message in the same class as the simple message, so I just created an entirely new class for the save message. I'm still having problems with it. Like, it won't update the login section for some reason, even though it should, but I feel like that'll be easier to fix.